% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{date.2dec}
\alias{date.2dec}
\title{Converts date(s) in Date format to date(s) in DYM's decimal dates. Useful for generating a tvect to be written into the DYM file.}
\usage{
date.2dec(dates)
}
\arguments{
\item{dates}{a date or a vector of dates in Date format.}
}
\value{
corresponding vector with its elements in decimal format.
}
\description{
Converts date(s) in Date format to date(s) in DYM's decimal dates. Useful for generating a tvect to be written into the DYM file.
}
\examples{
tvect <- date.2dec(as.Date("1975-01-15")); #returns 1975.038
}
