% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dec.2date}
\alias{dec.2date}
\title{Converts DYM's decimal date to the Date format}
\usage{
dec.2date(dec.date)
}
\arguments{
\item{dec.date}{is the date or an array of dates in decimal format as used in DYM files, i.e., year+(month-1)/12+(day_of_month-1)/366.}
}
\value{
corresponding date(s) in the Date format.
}
\description{
Converts DYM's decimal date to the Date format
}
\examples{
dec.2date(1975.0); #returns "1974-12-31"
dec.2date(1975.01); #returns "1975-01-03"
}
