% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DYM.R
\docType{package}
\name{DYM}
\alias{DYM}
\alias{DYM-package}
\title{DYM}
\usage{
DYM()
}
\description{
You might mistype an object name.
The package suggests the correct spell of the object you meant.

If the function is called after an error of 'object not found',
the function tries to tell you the name of the correct name that you meant.
}

