% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findSimilarName.R
\name{findSimilarName}
\alias{findSimilarName}
\title{Looks for approximate matches to \code{x} (the first argument)
within \code{name} (the second) argument.}
\usage{
findSimilarName(x, names, threshold = 2)
}
\arguments{
\item{x}{The name to search}

\item{names}{The candidates of the correct.}

\item{threshold}{The maximum distance between the misspell (\code{x}) and the correct answer (in \code{name}).}
}
\description{
Looks for approximate matches to \code{x} (the first argument)
within \code{name} (the second) argument.
}
\seealso{
agrep
}

