% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.ts.dym}
\alias{get.ts.dym}
\title{Aggregating DYM variable through space and returning time series}
\usage{
get.ts.dym(dym.in, ts = NULL, xs = NULL, ys = NULL, FUN = "sum", ivalue = NULL)
}
\arguments{
\item{dym.in}{a string containing the full name (with full path or relative to home directory) of the DYM file.}

\item{ts}{the set of two dates in Date format c(t1,t2) for subsetting. If left undefined, the original time period will be used.}

\item{xs}{the set of longitudes c(x1,x2) for subsetting. If lest undefined, the original longitudinal range will be used.}

\item{ys}{the set of latitudes in format c(y1,y2) for subsetting. If left undefined, the original latitudinal range will be used.}

\item{FUN}{a name of a function for aggregation, can take values "sum", "mean", "wmean", "mmean" for a sum, a simple mean, weighted mean and mass mean respectively.}

\item{ivalue}{invalid value to be used to ignore data.}
}
\value{
Returns a list with two array variables, t being the time vector and y being the aggregated variable.
}
\description{
Function reads DYM file, computes the sum (default) or the average over selected region and returns the time series. The sum always implies conversion to the mass content and sensible only for biomass densities. For averaging, it can return a simple average without modifying the units of the DYM data, or convert the data to the mass content (e.g., for biomass densities) by multiplying by the cell area before averaging, or a weighted average with weights being the grid cell areas (e.g., for temperature).
}
\examples{
dym.in <- system.file("extdata", "skj.dym", package = "dym")
txy <- read.txy.dym(dym.in)
txy
x.range <- c(100,180)
y.range <- c(-25,25)
dat <- get.ts.dym(dym.in,xs=x.range,ys=y.range); 
}
