% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissingVariable.R
\name{getMissingVariable}
\alias{getMissingVariable}
\title{Finds the misspelled object.}
\usage{
getMissingVariable()
}
\description{
When this function is called after an error,
it looks for the error message of missing value
and returns the name of the mistype if it is found.
}

