% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNames.R
\name{isVariableName}
\alias{isVariableName}
\title{Checks if the given \code{name} is valid as a variable name for R.}
\usage{
isVariableName(name)
}
\arguments{
\item{name}{A character vector to check.}
}
\description{
Checks if the given \code{name} is valid as a variable name for R.
}

