% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mdate.2dec}
\alias{mdate.2dec}
\title{Converts a vector of monthly dates to a vector of DYM's decimal dates. Useful when working with 360-days calendar to generate a tvect to be written into the DYM file.}
\usage{
mdate.2dec(dates)
}
\arguments{
\item{dates}{a vector with monthly dates as used for the 360-days calendar.}
}
\value{
corresponding vector with its elements in decimal format.
}
\description{
Converts a vector of monthly dates to a vector of DYM's decimal dates. Useful when working with 360-days calendar to generate a tvect to be written into the DYM file.
}
\examples{
dates <- gen.monthly.dates(as.Date("1970-01-15"),as.Date("1970-05-15"))
tvect <- mdate.2dec(dates);
}
