% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mult.dyms}
\alias{mult.dyms}
\title{Computing the product of two or more DYM files}
\usage{
mult.dyms(dyms.in, dym.out = NULL, out = FALSE)
}
\arguments{
\item{dyms.in}{the vector of strings containing the full names (with full path or relative to home directory) of the input DYM files.}

\item{dym.out}{the name of the output DYM file. If not specified, then the result will be returned (see out parameter).}

\item{out}{the logical value (FALSE by default) telling if the result should be returned by the function. It will be automatically turned to TRUE if dym.out is not provided.}

\item{average}{a logical indicating whether the result should be an average.}
}
\value{
Optionally, returns a list with dimensional variables, t, x and y, as well as z, a 3d variable.
}
\description{
Sum up multiple DYM files with identical dimensions, writing the result to new DYM file and/or returning it.
}
\examples{
Example 1. Get the early larvae that are accessible to the sampling gear by multiplying the density by the MLD function as used in SEAPODYM
dir.in  <- paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-jra55np_2x30d-po/")
dym.1 <- list.files(dir.in,pattern="mld",full.names=TRUE)
dym.out <- gsub("early","qmld_early_larvae",dym.2)
mult.dyms(c(dym.1,dim.2),dym.out); # writing the result to DYM file

}
