% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.restart.dym}
\alias{read.restart.dym}
\title{Reads restart DYM file}
\usage{
read.restart.dym(file.in)
}
\arguments{
\item{file.in}{the name of the DYM file.}
}
\description{
Reads DYM file with population density initial conditions
}
\examples{
dym <- read.restart.dym("skj_cohorts.dym");
summary(dym)
}
