% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.txy.dym}
\alias{read.txy.dym}
\title{Reads dimensional vectors from the DYM file}
\usage{
read.txy.dym(file.in, verbose = FALSE)
}
\arguments{
\item{file.in}{the name of the DYM file.}
}
\value{
A list with t, x and y, where time vector returned in decimal format, as written in the DYM file, x and y are the longitudinal and latitudinal vectors respectively.
}
\description{
Reads dimensional vectors from the DYM file
}
\examples{
dym.in <- system.file("extdata", "skj.dym", package = "dym");
txy <- read.txy.dym(dym.in)
summary(txy)
}
