% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sum.dyms}
\alias{sum.dyms}
\alias{ave.dyms}
\title{Computing the sum of multiple DYM files}
\usage{
\method{sum}{dyms}(dyms.in, dym.out = NULL, out = FALSE, average = FALSE)

ave.dyms(dyms.in, dym.out = NULL, out = FALSE)
}
\arguments{
\item{dyms.in}{the vector of strings containing the full names (with full path or relative to home directory) of the input DYM files.}

\item{dym.out}{the name of the output DYM file. If not specified, then the result will be returned (see out parameter).}

\item{out}{the logical value (FALSE by default) telling if the result should be returned by the function. It will be automatically turned to TRUE if dym.out is not provided.}

\item{average}{a logical indicating whether the result should be an average.}
}
\value{
Optionally, returns a list with dimensional variables, t, x and y, as well as z, a 3d variable.
}
\description{
Sum up multiple DYM files with identical dimensions, writing the result to new DYM file and/or returning it.
}
\section{Functions}{
\itemize{
\item \code{ave.dyms()}: Average multiple DYM files with identical dimensions, writing the results to new DYM file and/or returning it.

}}
\examples{
# Example 1. Get the epipelagic forage at night
dir.in <- paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-era5np_1x30d-po/forage/")
dyms.in <- paste0(dir.in,"Fbiom_",c("epi","mmeso","hmbathy"),".dym")
dym.out <- gsub("epi","epi_night",dyms.in[1])
sum.dyms(dyms.in,dym.out); # writing the result to DYM file

# Example 2. Get the density of recruits to be compared with MFCL class
setwd(paste0(Sys.getenv("SKJ_HOME"),"/REF/output/"))
dyms.in <- paste0("skj_age",4:6,".dym")
dym.out <- "skj_recruits.dym"
sum.dyms(dyms.in,dym.out); # writing the result to the DYM file
out <- sum.dyms(dyms.in,dym.out,out=TRUE); # writing DYM file and output the result
out <- ave.dyms(dyms.in); # output the average density in selected age classes
}
