% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{diff.dyms}
\alias{diff.dyms}
\title{Computing the difference between two DYM files}
\usage{
\method{diff}{dyms}(dym1, dym2, dym.out = NULL, out = FALSE)
}
\arguments{
\item{dym.out}{a string giving the name of output DYM file.}

\item{dym.1}{a string containing the full name (with full path or relative to home directory) of the first DYM file.}

\item{dym.2}{a string containing the full name (with full path or relative to home directory) of the second DYM file.}
}
\value{
Optionally, returns a list with dimensional variables, t, x and y, as well as z, a 3d variable.
}
\description{
Function computes the difference between two DYM files 'dym1-dym2' and either writes it to another DYM file or returns the result.
}
\examples{
setwd(paste0(Sys.getenv("BET_HOME"),"/REF/output/output_F0/"))
dym1 <- "bet_age25.dym"
dym2 <- "bet_age36.dym"
dym.out <- "bet_natloss_year.dym"
diff.dyms(dym1,dym2,dym.out);
out <- diff.dyms(dym1, dym2, dym.out, out=TRUE);
}
