% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{eapply.dyms}
\alias{eapply.dyms}
\title{Element-wise operation on DYM file(s) variables}
\usage{
eapply.dyms(dyms.in, dyms.out, FUN, ...)
}
\arguments{
\item{dyms.in}{a string or a vector of strings containing the full names (with full path or relative to home directory) of the input DYM files.}

\item{dyms.out}{a string or a vector of strings giving the full names of output DYM file(s). Must be provided, otherwise will do nothing.}

\item{FUN}{a function to be applied to the DYM variable on an element level, should be defined elsewhere or directly as an argument.}
}
\description{
Function applies a function to every element of DYM file(s) and writes the result to dyms.out with the same dimensions.
}
\examples{
dir.in  <- paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-era5-np_1x30d-glo/")
#1. multiply all elements of micronekton input data by a factor of 2:
dir.out <- paste0(dir.in,"forage_x",mult,"/")
if (!dir.exists) dir.create(dir.out)
dyms.in  <- list.files(paste0(dir.in,"forage/"),pattern="Fbiom",full.names=TRUE)
dyms.out <- paste0(dir.out,basename(dyms.in))
eapply.dyms(dyms.in,dyms.out,function(x)2*x);
#2. compute the gaussian function of sea surface temperature
dym.in  <- list.files(dir.in,pattern="sst",full.names=TRUE)
dym.out <- gsub("sst","gauss-sst",dym.in)
gauss <- function(x,a,b) exp(-((x-a)^2)/(2*b*b))
eapply.dyms(dym.in,dym.out,"gauss",a=29,b=2))   
}
