% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSimilarName.R
\name{findSimilarName}
\alias{findSimilarName}
\title{Looks for approximate matches to \code{x} (the first argument)
within \code{name} (the second) argument.}
\usage{
findSimilarName(x, names, threshold = 2, max_n = 10)
}
\arguments{
\item{x}{A string giving the (misspelt) name to search for.}

\item{names}{A character vector of correct names to match against.}

\item{threshold}{The maximum distance between the misspell (\code{x}) and the correct answer (in \code{name}).}

\item{max_n}{An integer limiting the number of results.  Passed to \code{\link[utils]{head}}.}
}
\description{
Looks for approximate matches to \code{x} (the first argument)
within \code{name} (the second) argument.
}
\examples{
\donttest{
x <- "logg"
names <- DYM:::getNames(x)
lapply(
  stats::setNames(0:4, 0:4), 
  function(i) DYM:::findSimilarName(x, names, threshold = i)
)
}
}
\seealso{
\code{\link[utils]{adist}} calculates the distance between strings.
\code{\link[base]{agrep}} and \code{\link[stringdist]{stringdist-package}} 
provide alternate metrics for these distances.
}

