% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.2d.dym}
\alias{get.2d.dym}
\title{Aggregating DYM variable through time and returning 2d variable}
\usage{
get.2d.dym(
  dym.in,
  dym.out = NULL,
  ts = NULL,
  xs = NULL,
  ys = NULL,
  FUN = "mean",
  ivalue = NULL,
  out = FALSE
)
}
\arguments{
\item{dym.in}{a string containing the full name (with full path or relative to home directory) of the DYM file.}

\item{dym.out}{a string giving the name of output DYM file if need to be written. In this case, it will be written in the DYM1 format for visualization in SeapodymView.}

\item{ts}{the set of two dates in Date format c(t1,t2) for subsetting. If left undefined, the original time period will be used.}

\item{xs}{the set of longitudes c(x1,x2) for subsetting. If lest undefined, the original longitudinal range will be used.}

\item{ys}{the set of latitudes in format c(y1,y2) for subsetting. If left undefined, the original latitudinal range will be used.}

\item{FUN}{a function to be used for aggregation, e.g., "sum" or "mean" (default).
ivalue invalid value to be used to ignore data.}

\item{out}{the logical value (FALSE by default) telling if the output should be returned by the function. It will be automatically turned to TRUE if dym.out is not provided.}
}
\description{
Function reads DYM file, computes the average (default) or the sum over selected time period and returns the result as a 2d array and/or writes it to a DYM1 file. Note, the only purpose of writing DYM1 file is to visualize it in SeapodymView.
}
\examples{
Example 1. Subsetting and computing average decadal distribution, returning the result.
setwd(paste0(Sys.getenv("SKJ_HOME"),"/REF/output/"))
dym.in  <- "skj_Ha_first_maturity.dym"
t.range <- c(as.Date("1981-1-1"),as.Date("1990-12-31"))
x.range <- c(90,290)
y.range <- c(-40,45)
dat <- get.2d.dym(dym.in,ts=t.range,xs=x.range,ys=y.range); 
Example 2. Average primary production, removing cells with invalid values.
setwd(paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-era5-np_1x30d-glo/"))
dym.in <- list.files(pattern="_pp_")
dat <- get.2d.dym(dym.in,ts=t.range,xs=x.range,ys=y.range,ivalue=-999);
require(seamaps)
image(dat$x,dat$y,dat$z,xlab='',ylab='',xaxt='n',yaxt='n',main="Primary production, 1981-1990")
nice.map.over(x.range,y.range)
}
