% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getNames.R
\name{getNames}
\alias{getNames}
\title{Retrieves available symbols in the specified environment.}
\usage{
getNames(envir = .GlobalEnv)
}
\arguments{
\item{envir}{The base environment to search variables.}
}
\description{
Retrieves available symbols in the specified environment.
}

