% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNames.R
\name{getNames}
\alias{getNames}
\title{Retrieves available symbols in the specified environment.}
\usage{
getNames(mode, envir = .GlobalEnv)
}
\arguments{
\item{mode}{The mode of misspelled name.}

\item{envir}{The base environment to search variables.}
}
\description{
Retrieves available symbols in the specified environment.
}

