% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{make.climatology}
\alias{make.climatology}
\title{Making climatology of a DYM variable}
\usage{
make.climatology(dym.in, imonths = 1:12, ts = NULL, out = FALSE)
}
\arguments{
\item{dym.in}{a string containing the full name (with full path or relative to home directory) of the DYM file.}

\item{imonths}{the set of first months indices for averaging. For example, 1:12 corresponds to monthly climatology, c(1,3,7,10) to quarterly. Note, the last month of each climatological interval is always the month preceding the first month of the next, or the first element in a set.}

\item{ts}{the set of two dates in Date format c(t1,t2) for subsetting. If left undefined, the original time period will be used.}

\item{out}{the logical value (FALSE by default) telling if the output should be returned by the function, in addition to be writted to the DYM file.}
}
\value{
Optionally, returns a list with dimensional variables, x and y, as well as z, a climatological 3d variable.
}
\description{
Function reads DYM file and generates the climatology (average) over selected time period and months. The output DYM file will be written in the same directory and name of the input DYM file with added suffix 'clm_' and either 'monthly' (default), 'quarterly', 'seasonal' or 'other', depending on the selection of months.
}
\examples{
dym.in   <- paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-era5-np_1x30d-glo/forage/Fbiom_epi.dym")
seasons  <- c(12,3,6,9)  # making climatology over calendar seasons: dec-feb, mar-may, un-aug, sep-nov.
dat <- gen.climatology(dym.in,seasons,out=TRUE); # writing to the DYM file Fbiom_epi_clm_seasonal.dym and returning result
}
