% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mult.dyms}
\alias{mult.dyms}
\title{Computing the product of two or more DYM files}
\usage{
mult.dyms(dyms.in, dym.out = NULL, out = FALSE)
}
\arguments{
\item{dyms.in}{the vector of strings containing the full names (with full path or relative to home directory) of the input DYM files.}

\item{dym.out}{the name of the output DYM file. If not specified, then the result will be returned (see out parameter).}

\item{out}{the logical value (FALSE by default) telling if the result should be returned by the function. It will be automatically turned to TRUE if dym.out is not provided.}

\item{average}{a logical indicating whether the result should be an average.}
}
\value{
Optionally, returns a list with dimensional variables, t, x and y, as well as z, a 3d variable.
}
\description{
Sum up multiple DYM files with identical dimensions, writing the result to new DYM file and/or returning it.
}
\examples{
# Example 1. Get the skipjack denstiy that is accessible to the fishing gear that can fish only in waters with euphotic depth < 100 m. SEAPODYM environment is needed to run this example.
dir.in  <- paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-interim_2x30d_po/")
dym.1 <- list.files(dir.in,pattern="zeu",full.names=TRUE)
dym.2 <- system.file("extdata", "skj.dym", package = "dym")
txy <- read.txy.dym(dym.2)
dym.1s <- paste0("my-",basename(dym.1))
subset.dyms(dym.1,dym.1s,range(dec.2date(txy$t)),range(txy$x),range(txy$y))
qfun <- function(x,slope,depth) 1/(1+exp((slope)*(x-depth)))
dym.1sq <- paste0("my-q-",basename(dym.1s))
eapply.dyms(dym.1s,dym.1sq,FUN="qfun",slope=0.5,depth=100)
dym.out <- paste0("accessible-",basename(dym.2))
mult.dyms(c(dym.1sq,dym.2),dym.out); # writing the result to dym.out file

}
