% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.var.dym}
\alias{read.var.dym}
\title{Reading DYM file}
\usage{
read.var.dym(
  file.in,
  t0.user = NULL,
  tfin.user,
  region = c(NA, NA, NA, NA),
  dt = 30,
  apply.mask = FALSE,
  t.date.format = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file.in}{the DYM file accessible with provided path}

\item{t0.user}{the date provided as c(year,month,day) to be used the first date for extraction. If not specified, start from the first time step written in the DYM file.}

\item{tfin.user}{the date provided as c(year,month,day) to be used the last date for extraction. If not specified, stop at last time step written in the DYM file.}

\item{region}{the set of four values c(x1,x2,y1,y2) defining east-west and south-north coordinates respectively.}

\item{verbose}{flag (default value is TRUE) controlling the prompt for the nominal function execution.}
}
\value{
The list of variables written in the DYM file, that is a vector of longitudes, x, latitudes, y, time steps, t, the land mask matrix, landmask and the data 3d matrices, var[t,x,y].
}
\description{
Reads 3d variable from the DYM file, which has dimensions (time,longitude,latitude), and returns the list with time vector, coordinates, land mask and variable data.
}
\examples{
data <- read.var.dym("sst.dym");
var<-data$var; tt<-data$t; x<-data$x; y<-data$y
}
