% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{subset.dyms}
\alias{subset.dyms}
\title{Subsetting one or multiple DYM files}
\usage{
\method{subset}{dyms}(dyms.in, dyms.out = NULL, ts = NULL, xs = NULL, ys = NULL)
}
\arguments{
\item{dyms.in}{the vector of strings containing the full names (with full path or relative to home directory) of the input DYM files.}

\item{dyms.out}{the name of the output DYM file. If not specified, then the same names will be used for the subsetted files.}

\item{ts}{the set of two dates in Date format c(t1,t2) for subsetting. If left undefined, the original time period will be used.}

\item{xs}{the set of longitudes c(x1,x2) for subsetting. If lest undefined, the original longitudinal range will be used.}

\item{ys}{the set of latitudes in format c(y1,y2) for subsetting. If left undefined, the original latitudinal range will be used.}
}
\description{
Function subsets multiple DYM files with identical dimensions. It will subset each file and write the new set of DYMs with the same names, unless the new names are specified, into the following folder 'subset-year1-year2_x1-x2_y1-y2', where dim1 is the first and dim2 is the last dimensional variable.
}
\examples{
t.range <- c(as.Date("1980-1-1"),as.Date("2020-12-31"))
x.range <- c(100,180)
y.range <- c(-50,50)
dyms.in <- list.files(paste0(Sys.getenv("SEA_FORCING_HOME"),"/run-era5-np_1x30d-glo"),full.names=TRUE,pattern=".dym")
subset.dyms(dyms.in,ts=t.range,xs=x.range,ys=y.range);
}
